Access message descriptors through `UIMessageStore`
This is so that message descriptors can be modified in such a way
that the change will affest the whole app and so that UI will
appropriately re-render reactively when message content is modified.

When message descriptors were imported directly in multiple components,
modifying one imported instance did not seem to be able to affect
the others.